+++ src/module_ntp.cpp
@@ -9,7 +9,7 @@
 #include <ArduinoJson.h>
 #include "FSWebServerLib.h"
 #include "debug.h"
-#include "wifi_mod.h"
+#include "module_wifi.h"
 #include "module_udp.h"
 #include "module_ntp.h"
